+++ mock/response.go
@@ -15,11 +15,11 @@ func NewResponse() porthos.Response {
 type Response struct {
 	Body        []byte
 	ContentType string
-	StatusCode  int16
+	StatusCode  int32
 	Headers     *porthos.Headers
 }
 
-func (r *Response) JSON(statusCode int16, Body interface{}) {
+func (r *Response) JSON(statusCode int32, Body interface{}) {
 	jsonBody, err := json.Marshal(&Body)
 
 	if err != nil {
@@ -31,13 +31,13 @@ func (r *Response) JSON(statusCode int16, Body interface{}) {
 	r.ContentType = "application/json"
 }
 
-func (r *Response) Raw(statusCode int16, contentType string, body []byte) {
+func (r *Response) Raw(statusCode int32, contentType string, body []byte) {
 	r.StatusCode = statusCode
 	r.Body = body
 	r.ContentType = contentType
 }
 
-func (r *Response) Empty(statusCode int16) {
+func (r *Response) Empty(statusCode int32) {
 	r.StatusCode = statusCode
 }
 
@@ -45,7 +45,7 @@ func (r *Response) GetHeaders() *porthos.Headers {
 	return r.Headers
 }
 
-func (r *Response) GetStatusCode() int16 {
+func (r *Response) GetStatusCode() int32 {
 	return r.StatusCode
 }
